+++ Source/Core/Common/Src/Thread.cpp
@@ -356,7 +356,8 @@ namespace Common
 		if (ret) ERROR_LOG(COMMON, "%s: pthread_create(%p, %p, %p, %p) failed: %s\n", 
 						   __FUNCTION__, &thread_id, &attr, function, arg, strerror(ret));
 		
-		INFO_LOG(COMMON, "created new thread %lu (func=%p, arg=%p)\n", thread_id, function, arg);
+		INFO_LOG(COMMON, "created new thread %lu (func=%p, arg=%p)\n",
+			(unsigned long)thread_id, function, arg);
 	}
 	
 	
@@ -372,9 +373,13 @@ namespace Common
 		{
 			void* exit_status;
 			int ret = pthread_join(thread_id, &exit_status);
-			if (ret) ERROR_LOG(COMMON, "error joining thread %lu: %s\n", thread_id, strerror(ret));
+			if (ret) ERROR_LOG(COMMON,
+				"error joining thread %lu: %s\n",
+				(unsigned long)thread_id, strerror(ret));
 			if (exit_status)
-				ERROR_LOG(COMMON, "thread %lu exited with status %d\n", thread_id, *(int *)exit_status);
+				ERROR_LOG(COMMON,
+				"thread %lu exited with status %d\n",
+				(unsigned long)thread_id, *(int *)exit_status);
 			thread_id = 0;
 		}
 	}
